Consolidate Bing Ads into single folder - merge bing-ads-v1 into bing-ads, update tool/block URLs
