--- apps/sim/app/api/bing-ads-v1/query/index.ts
@@ -1,12 +1,0 @@
-/**
- * Bing Ads V1 API - Exports
- * Following Google Ads v1 pattern
- */
-
-export * from './ai-provider'
-export * from './bing-api'
-export * from './constants'
-export * from './query-generation-simple'
-export * from './result-processing'
-export * from './route'
-export * from './types'
