+++ apps/sim/app/api/bing-ads/accounts/route.ts
