+++ apps/sim/app/api/bing-ads/query/prompt.ts
